yolo/requirements.txt
albumentations
comet
Flask
Flask-MQTT
Flask-UUID
gsutil
gunicorn
matplotlib
notebook
numpy
opencv-python<4.6.0.66
pandas
Pillow
psutil
PyYAML
requests
seaborn
tqdm
